+++ src/components/types/reviewType.ts
@@ -25,3 +25,22 @@ export interface FilterState {
   date: Date | null;
   sort: { id: number; name: string } | null;
 }
+
+export interface CroppedImageType {
+  objectURL: string;
+  blobImg: Blob | null;
+}
+
+export interface ReviewFormValues {
+  score: number;
+  comment: string;
+  images: File[];
+  fileUrls?: string[];
+}
+
+export interface ReviewModalProps {
+  mode: 'create' | 'edit';
+  initialData?: { score: number; comment: string; images?: string[] };
+  onSubmit: (data: ReviewFormValues) => void;
+  onClose: () => void;
+}
